+++ src/FileServer.ts
@@ -1,4 +1,5 @@
 // Example: FileServer.ts
+// Purpose: Basic HTTP server to serve static files. The HTTP content type is set according to the file name extension.
 
 import { ABHttpServer, ABRequest } from './ABHttpServer'
 import { ServerResponse } from 'http';