+++ src/GetPost.js
@@ -1,5 +1,10 @@
 "use strict";
 // Example: GetPost.ts
+// Purpose: Start a HTTP and HTTPS server on ports 8080 and 8081. All requests from both servers are
+//          handled thru the same class methods. For GET requests with an URL / stats, the server
+//          sends back a JSON object with some server statistic, all other GET requests will receive
+//          a text line with the send URL.All POST requests will be answered with the data sent in the
+//          body of the POST request.
 var __extends = (this && this.__extends) || (function () {
     var extendStatics = function (d, b) {
         extendStatics = Object.setPrototypeOf ||