+++ src/HelloWorld.js
@@ -1,5 +1,6 @@
 "use strict";
 // Example: HelloWorld.ts
+// Purpose: Sends back the response "Hello World!" to every client request
 var __extends = (this && this.__extends) || (function () {
     var extendStatics = function (d, b) {
         extendStatics = Object.setPrototypeOf ||