+++ src/HelloWorld.ts
@@ -1,4 +1,5 @@
 // Example: HelloWorld.ts
+// Purpose: Sends back the response "Hello World!" to every client request
 
 import { ABHttpServer, ABRequest } from './ABHttpServer'
 import { ServerResponse } from 'http';