+++ src/ShowCookie.ts
@@ -6,7 +6,7 @@ import { ServerResponse } from 'http';
 
 class MyServer extends ABHttpServer {
   get(request: ABRequest, response: ServerResponse) {
-    this.sendText(response, `Unique session identifier: ${request.sessionId}`)
+    this.sendText(response, `The unique session identifier is ${request.sessionId}`)
   }
 }
 var myServer = new MyServer(8080)